refactor: 'ui' 폴더 위치 'components' 폴더 하위로 이동
