+++ src/components/layout/Header/Header.jsx
@@ -3,8 +3,8 @@ import { css } from "@emotion/react";
 import logoPandaImg from "../../assets/images/logo-panda.svg";
 import logoTxtImg from "../../assets/images/logo-txt.svg";
 import avatarImg from "../../assets/images/img-avatar.png";
-import Nav from "../Nav";
-import Avatar from "../Avatar";
+import Nav from "../../Nav";
+import Avatar from "../../Avatar";
 
 const Header = () => {
   const handleAvatarClick = () => {
